+++ addons/uh60_fms/config/keypad.hpp
@@ -59,13 +59,6 @@ class FMS_Keypad {
     radius = 0.025;
     buttonUp = [(_this select 0),[ARR_2(FMS_PAGE_INDEX,FMS_PAGE_PERFORMANCE)], true] call vtx_uh60_fms_fnc_interaction_pageChange;
   };
-  class FMS_BTN_COM {
-    positionType = "coordinates";
-    position[] = {0.0688477,4.81626,-0.629442};
-    label = "COM";
-    radius = 0.025;
-    buttonUp = [(_this select 0),[ARR_2(FMS_PAGE_INDEX,FMS_PAGE_COMM_ACRE)], true] call vtx_uh60_fms_fnc_interaction_pageChange;
-  };
   class FMS_BTN_NAV {
     positionType = "coordinates";
     position[] = {0.0810547,4.81627,-0.629442};
