+++ addons/uh60_fms/config/keypadCopilot.hpp
@@ -59,13 +59,6 @@ class FMS_Keypad {
     radius = 0.025;
     buttonUp = [(_this select 0),[ARR_2(FMS_PAGE_INDEX,FMS_PAGE_PERFORMANCE)], true] call vtx_uh60_fms_fnc_interaction_pageChange;
   };
-  class FMS_BTN_COM {
-    positionType = "coordinates";
-    position[] = {-0.1992183,4.81418,-0.629442};
-    label = "COM";
-    radius = 0.025;
-    buttonUp = [(_this select 0),[ARR_2(FMS_PAGE_INDEX,FMS_PAGE_COMM_ACRE)], true] call vtx_uh60_fms_fnc_interaction_pageChange;
-  };
   class FMS_BTN_NAV {
     positionType = "coordinates";
     position[] = {-0.1870113,4.81419,-0.629442};
